refactor: retire ingest utils
