+++ src/handlers/http/modal/utils/mod.rs
@@ -16,6 +16,5 @@
  *
  */
 
-pub mod ingest_utils;
 pub mod logstream_utils;
 pub mod rbac_utils;
